refactor: divided Calendar component into smaller components
