+++ frontend/src/components/calendar/index.js
@@ -1,0 +1,7 @@
+export { default as CalendarHeader } from './CalendarHeader';
+export { default as DayView } from './DayView';
+export { default as EventForm } from './EventForm';
+export { default as EventItem } from './EventItem';
+export { default as EventLegend } from './EventLegend';
+export { default as MonthView } from './MonthView';
+export { default as Timeline } from './Timeline';
